+++ system/src/migrations/updates/_module.mjs
@@ -1,2 +1,4 @@
 export {default as Update_240218_1} from "./Update_240218_1.mjs";
 export {default as Update_240526_1} from "./Update_240526_1.mjs";
+export {default as Update_240526_2} from "./Update_240526_2.mjs";
+export {default as Update_240527_1} from "./Update_240527_1.mjs";