+++ system/src/roller/Dialog2d20.mjs
@@ -15,16 +15,18 @@ export default class Dialog2d20 extends Dialog {
 		options={}
 	) {
 		super(dialogData, options);
-		this.rollName = rollName;
-		this.diceNum = diceNum;
-		this.attribute = attribute;
-		this.skill = skill;
-		this.focus = focus;
-		this.complication = complication;
+
 		this.actor = actor;
-		this.prefAttribute = prefAttribute;
 		this.actorId = actorId;
+		this.attribute = attribute;
+		this.complication = complication;
+		this.diceNum = diceNum;
+		this.focus = focus;
 		this.itemId = itemId;
+		this.prefAttribute = prefAttribute;
+		this.rollName = rollName;
+		this.skill = skill;
+
 		this.options.classes = ["dice-icon"];
 	}
 
@@ -42,23 +44,26 @@ export default class Dialog2d20 extends Dialog {
 		});
 
 		html.on("click", ".roll", event => {
-			let attr = html.find('[name="attribute"]').val();
-			if (!attr) {
-				let attrAbr = html.find(".select-attribute").val();
-				attr = this.actor.attributes[attrAbr].value;
+			let attribute = html.find('[name="attribute"]').val();
+
+			if (!attribute) {
+				let attributeShortName = html.find(".select-attribute").val();
+				attribute = this.actor.attributes[attributeShortName].value;
 			}
-			let skill = html.find('[name="skill"]').val();
-			let complication = html.find('[name="complication"]').val();
-			let isFocus = html.find('[name="focus"]').is(":checked");
+
+			const skill = html.find('[name="skill"]').val();
+			const complication = html.find('[name="complication"]').val();
+			const isFocus = html.find('[name="focus"]').is(":checked");
+
 			game.ac2d20.Roller2D20.rollD20({
-				rollname: this.rollName,
-				dicenum: this.diceNum,
-				attribute: attr,
-				skill: skill,
-				focus: isFocus,
-				complication: complication,
 				actorId: this.actorId,
+				attribute,
+				complication,
+				dicenum: this.diceNum,
+				focus: isFocus,
 				itemId: this.itemId,
+				rollName: this.rollName,
+				skill,
 			});
 		});
 	}
@@ -68,7 +73,18 @@ export default class Dialog2d20 extends Dialog {
 		$(html).find(`[data-index="${this.diceNum}"]`).addClass("marked");
 	}
 
-	static async createDialog({ rollName = "Roll D20", diceNum = 2, attribute = 0, skill = 0, focus = false, complication = 20, actor = null, prefAttribute = null, actorId = null, itemId=null } = {}) {
+	static async createDialog({
+		actor = null,
+		actorId = null,
+		attribute = 0,
+		complication = 20,
+		diceNum = 2,
+		focus = false,
+		itemId = null,
+		prefAttribute = null,
+		rollName = "Roll D20",
+		skill = 0,
+	} = {}) {
 		let dialogData = {};
 		dialogData.rollName = rollName;
 		dialogData.diceNum = diceNum;
@@ -81,7 +97,12 @@ export default class Dialog2d20 extends Dialog {
 		dialogData.prefAttribute = prefAttribute;
 		dialogData.actorId = actorId;
 		dialogData.itemId = itemId;
-		const html = await renderTemplate("systems/ac2d20/templates/dialogs/dialog2d20.hbs", dialogData);
+
+		const html = await renderTemplate(
+			"systems/ac2d20/templates/dialogs/dialog2d20.hbs",
+			dialogData
+		);
+
 		let d = new Dialog2d20(
 			rollName,
 			diceNum,